Remove SolutionExplorer ProjectItem classes added in 56136680f
