+++ src/VisualStudio/Core/SolutionExplorerShim/AnalyzersFolderItem/AnalyzersFolderItem.cs
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.Internal.VisualStudio.PlatformUI;
 using Microsoft.VisualStudio.Imaging;
@@ -17,6 +18,7 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
         private readonly ProjectId _projectId;
         private readonly IVsHierarchyItem _parentItem;
         private readonly IContextMenuController _contextMenuController;
+
 
         public AnalyzersFolderItem(
             Workspace workspace,
@@ -70,11 +72,6 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
         public override IContextMenuController ContextMenuController
         {
             get { return _contextMenuController; }
-        }
-
-        internal Project GetProject()
-        {
-            return _workspace.CurrentSolution.GetProject(_projectId);
         }
 
         /// <summary>
