+++ src/VisualStudio/Core/SolutionExplorerShim/AnalyzersFolderItem/AnalyzersFolderItemProvider.cs
@@ -4,8 +4,10 @@ using System;
 using System.Collections.Immutable;
 using System.ComponentModel.Composition;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using Microsoft.CodeAnalysis;
 using Microsoft.Internal.VisualStudio.PlatformUI;
+using Microsoft.VisualStudio.ComponentModelHost;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
 using Microsoft.VisualStudio.Utilities;
@@ -15,25 +17,29 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
     [Export(typeof(IAttachedCollectionSourceProvider))]
     [Name("AnalyzersFolderProvider")]
     [Order(Before = HierarchyItemsProviderNames.Contains)]
-    internal sealed class AnalyzersFolderItemProvider : ProjectFolderItemProvider
+    internal class AnalyzersFolderItemProvider : AttachedCollectionSourceProvider<IVsHierarchyItem>
     {
+        private readonly IComponentModel _componentModel;
         private readonly IAnalyzersCommandHandler _commandHandler;
+        private IHierarchyItemToProjectIdMap _projectMap;
+        private Workspace _workspace;
 
         [ImportingConstructor]
         public AnalyzersFolderItemProvider(
             [Import(typeof(SVsServiceProvider))] IServiceProvider serviceProvider,
-            [Import(typeof(AnalyzersCommandHandler))] IAnalyzersCommandHandler commandHandler) :
-            base(serviceProvider)
+            [Import(typeof(AnalyzersCommandHandler))] IAnalyzersCommandHandler commandHandler)
         {
+            _componentModel = (IComponentModel)serviceProvider.GetService(typeof(SComponentModel));
             _commandHandler = commandHandler;
         }
 
         /// <summary>
         /// Constructor for use only in unit tests. Bypasses MEF to set the project mapper, workspace and glyph service.
         /// </summary>
-        internal AnalyzersFolderItemProvider(IHierarchyItemToProjectIdMap projectMap, Workspace workspace, IAnalyzersCommandHandler commandHandler) :
-            base(projectMap, workspace)
+        internal AnalyzersFolderItemProvider(IHierarchyItemToProjectIdMap projectMap, Workspace workspace, IAnalyzersCommandHandler commandHandler)
         {
+            _projectMap = projectMap;
+            _workspace = workspace;
             _commandHandler = commandHandler;
         }
 
@@ -50,16 +56,41 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
                 var projectTreeCapabilities = GetProjectTreeCapabilities(hierarchy, itemId);
                 if (projectTreeCapabilities.Any(c => c.Equals("References")))
                 {
-                    var projectId = TryGetProject(item.Parent);
-                    if (projectId != null)
-                    {
-                        var workspace = TryGetWorkspace();
-                        return new AnalyzersFolderItemSource(workspace, projectId, item, _commandHandler);
-                    }
+                    return CreateCollectionSourceCore(item.Parent, item);
                 }
             }
 
             return null;
+        }
+
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private IAttachedCollectionSource CreateCollectionSourceCore(IVsHierarchyItem parentItem, IVsHierarchyItem item)
+        {
+            var hierarchyMapper = TryGetProjectMap();
+
+            ProjectId projectId;
+            if (hierarchyMapper != null &&
+                hierarchyMapper.TryGetProjectId(parentItem, out projectId))
+            {
+                var workspace = TryGetWorkspace();
+                return new AnalyzersFolderItemSource(workspace, projectId, item, _commandHandler);
+            }
+
+            return null;
+        }
+
+        private static ImmutableArray<string> GetProjectCapabilities(IVsHierarchy hierarchy)
+        {
+            object capabilitiesObj;
+            if (hierarchy.GetProperty((uint)VSConstants.VSITEMID.Root, (int)__VSHPROPID5.VSHPROPID_ProjectCapabilities, out capabilitiesObj) == VSConstants.S_OK)
+            {
+                var capabilitiesString = (string)capabilitiesObj;
+                return ImmutableArray.Create(capabilitiesString.Split(' '));
+            }
+            else
+            {
+                return ImmutableArray<string>.Empty;
+            }
         }
 
         private static ImmutableArray<string> GetProjectTreeCapabilities(IVsHierarchy hierarchy, uint itemId)
@@ -74,6 +105,36 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.SolutionExplore
             {
                 return ImmutableArray<string>.Empty;
             }
+        }
+
+        private Workspace TryGetWorkspace()
+        {
+            if (_workspace == null)
+            {
+                var provider = _componentModel.DefaultExportProvider.GetExportedValueOrDefault<ISolutionExplorerWorkspaceProvider>();
+                if (provider != null)
+                {
+                    _workspace = provider.GetWorkspace();
+                }
+            }
+
+            return _workspace;
+        }
+
+        private IHierarchyItemToProjectIdMap TryGetProjectMap()
+        {
+            var workspace = TryGetWorkspace();
+            if (workspace == null)
+            {
+                return null;
+            }
+
+            if (_projectMap == null)
+            {
+                _projectMap = workspace.Services.GetService<IHierarchyItemToProjectIdMap>();
+            }
+
+            return _projectMap;
         }
     }
 }
